pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "python_grpc_template"
authors = [{name = "Anthony Corletti", email = "python_grpc_template@example.com"}]
home-page = "https://www.github.com/anthonycorletti/python-grpc-template"
dynamic = ["version", "description"]
requires-python = ">=3.9.6"
dependencies = [
    "grpcio-tools>=1.44.0,<1.46.0",
    "grpcio>=1.44.0,<1.46.0",
]

[project.optional-dependencies]
test = [
    "pytest>=6.2.5,<7.0.0",
    "coverage>=6.1.1,<7.0.0",
    "pytest-cov>=3.0.0,<4.0.0",
    "grpcio-testing>=1.44.0,<1.46.0"
]
dev = [
    "mypy>=0.910,<1",
    "mypy-protobuf>=3.0.0,<4.0.0",
    "flake8>=3.9.2,<4.0.0",
    "black>=22,<23",
    "isort>=5.9.3,<6.0.0",
    "autoflake>=1.4,<2.0",
    "flake8-docstrings>=1.6.0,<2.0.0",
    "types-protobuf>=3.18.1,<4.0.0"
]

[tool.isort]
profile = "black"
